% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch38to19}
\alias{ch38to19}
\title{a liftover chain instance for hg38->hg19}
\format{An object of class \code{Chain} of length 25.}
\usage{
ch38to19
}
\description{
a liftover chain instance for hg38->hg19
}
\keyword{datasets}
