% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensRsLoc.R
\name{ensRsLoc}
\alias{ensRsLoc}
\title{use ensembl REST API to acquire SNP location for an rsid, lifting over as necessary}
\usage{
ensRsLoc(rsid = "rs6060535", ch = NULL, chk38 = FALSE,
  target_genome = "hg19")
}
\arguments{
\item{rsid}{character(1)}

\item{ch}{a liftOver chain object as imported with rtracklayer}

\item{chk38}{logical(1) to check that the query was resolved with assembly_name 'GRCh38'}

\item{target_genome}{character(1) just used for tag on result}
}
\description{
use ensembl REST API to acquire SNP location for an rsid, lifting over as necessary
}
\examples{
ensRsLoc()
}
