% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsd_vcf}
\alias{gsd_vcf}
\title{a CompressedVcf instance of 2504 genotypes at snps rs2305480, rs2305479 which are near GSDMB}
\format{An object of class \code{CollapsedVCF} with 2 rows and 2504 columns.}
\usage{
gsd_vcf
}
\description{
a CompressedVcf instance of 2504 genotypes at snps rs2305480, rs2305479 which are near GSDMB
}
\keyword{datasets}
