% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HG00096_rna_haps}
\alias{HG00096_rna_haps}
\title{a table of RNA-based haplotype calls derived from phaser}
\format{An object of class \code{data.frame} with 520 rows and 16 columns.}
\usage{
HG00096_rna_haps
}
\description{
a table of RNA-based haplotype calls derived from phaser
}
\keyword{datasets}
