% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look1kg.R
\name{look1kg}
\alias{look1kg}
\title{query 1000 genomes genotypes at pairs of SNP}
\usage{
look1kg(rsids = c("rs71503074", "rs71503075", "rs12084944"),
  ch = BiocRnaHap::ch38to19,
  src = "http://1000genomes.s3.amazonaws.com/release/20130502/ALL.\%s.phase3_shapeit2_mvncall_integrated_v5a.20130502.genotypes.vcf.gz",
  rsonly = TRUE, snvonly = TRUE)
}
\arguments{
\item{rsids}{character vector}

\item{ch}{chain object from rtracklayer}

\item{src}{character(1) sprintf template for 1000 genomes vcf, chr[nn] to be supplied for pct-s}

\item{rsonly}{logical(1), exclude 'esv' records}

\item{snvonly}{logical(1), exclude records with nchar(REF)>1}
}
\description{
query 1000 genomes genotypes at pairs of SNP
}
\note{
Typically SNPs are identified by rs number, but there are
some workstreams in which they come back as snp_cc_lllll where cc
is a chromosome identifier and lllll is a location.  We parse
the 'rsids' input for '^snp' and if we find this we skip
the step of resolving the rsid to genomic coordinates, using
the cc and lllll for location.
}
\examples{
lk2 = look1kg(c("rs71503074","rs71503075","rs12084944"))
table(geno(lk2)$GT[1,], geno(lk2)$GT[2,], geno(lk2)$GT[3,])
}
