% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NA06986_rnahaps}
\alias{NA06986_rnahaps}
\title{a table of RNA-based haplotype calls derived from phaser}
\format{An object of class \code{data.frame} with 69955 rows and 16 columns.}
\source{
2019 NHGRI Computational Medicine Hackathon, Martin Steinegger's computations
}
\usage{
NA06986_rnahaps
}
\description{
a table of RNA-based haplotype calls derived from phaser
}
\examples{
head(NA06986_rnahaps)
}
\keyword{datasets}
