% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RnaHapSet}
\alias{RnaHapSet}
\title{encapsulate a list of SNP configurations produced as haplotypes by phaser}
\usage{
RnaHapSet(genesym, radius, source, haplist = variant_options(genesym,
  radius))
}
\arguments{
\item{genesym}{character(1) symbol}

\item{radius}{numeric(1) region around gene to include}

\item{source}{a HapTab instance}

\item{haplist}{a list as returned by `variant_options`}
}
\description{
encapsulate a list of SNP configurations produced as haplotypes by phaser
}
\examples{
ht = HapTab(NA06986_rnahaps, "phaser NA06986")
RnaHapSet("ORMDL3", radius=50000, source=ht, 
    haplist=variant_options("ORMDL3", 50000))
}
