% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{variant_options}
\alias{variant_options}
\title{create list of snpids in RNA-defined haplotypes}
\usage{
variant_options(sym = "ORMDL3", radius = 50000,
  rna_hap_tab = NA06986_rnahaps, ...)
}
\arguments{
\item{sym}{character(1) gene symbol}

\item{radius}{numeric(1) interval around gene to include}

\item{rna_hap_tab}{a data.frame as generated by phaser}
}
\value{
a list
}
\description{
create list of snpids in RNA-defined haplotypes
}
\examples{
variant_options()
}
